fix: separado interface e itens em seus devidos arquivos
