+++ src/lib/interfaces.ts
@@ -1,0 +1,7 @@
+export interface Card {
+  id: string;
+  title: string;
+  description: string;
+  imageHref: string;
+  pageHref: string;
+}
