+++ src/lib/itens.ts
@@ -1,0 +1,55 @@
+import { v4 } from "uuid";
+
+import { Card } from "@/lib/interfaces";
+
+export const projectsCards: Card[] = [
+  {
+    id: v4(),
+    imageHref: "https://i.imgur.com/VyPhjnu.jpg",
+    description: "Descrição1",
+    title: "Titulo1",
+    pageHref: "/",
+  },
+  {
+    id: v4(),
+    imageHref: "https://i.imgur.com/VyPhjnu.jpg",
+    description: "Descrição2",
+    title: "Titulo2",
+    pageHref: "/",
+  },
+  {
+    id: v4(),
+    imageHref: "https://i.imgur.com/VyPhjnu.jpg",
+    description: "Descrição3",
+    title: "Titulo3",
+    pageHref: "/",
+  },
+  {
+    id: v4(),
+    imageHref: "https://i.imgur.com/VyPhjnu.jpg",
+    description: "Descrição3",
+    title: "Titulo3",
+    pageHref: "/",
+  },
+  {
+    id: v4(),
+    imageHref: "https://i.imgur.com/VyPhjnu.jpg",
+    description: "Descrição2",
+    title: "Titulo2",
+    pageHref: "/",
+  },
+  {
+    id: v4(),
+    imageHref: "https://i.imgur.com/VyPhjnu.jpg",
+    description: "Descrição3",
+    title: "Titulo3",
+    pageHref: "/",
+  },
+  {
+    id: v4(),
+    imageHref: "https://i.imgur.com/VyPhjnu.jpg",
+    description: "Descrição3",
+    title: "Titulo3",
+    pageHref: "/",
+  },
+];
